Move to workspaces, add fixphrase
This is more of a fiddle than a real functional improvement.  I added a
useless feature that is slightly more than a demo but forced some
structure onto the work.  This required a refactor and doing that
refactor now, before we have a lot of code, will prevent some pain
later.  So... yay useless feature but YAY working tests.