Don't mount static files when testing
When we test, we are one dir down, in the neems-core crate.  When we run
for real, we are in the top-level workspace dir.  If we try to pull in
the static stuff, we end up one dir level off and rocket complains that
the static dir is bogus.  We can fix that with a symlink, but we don't
actually test the static pages so let's just not mount them.