Set up CI
This sets up linting and test CI scripts so we can avoid breaking the
build some day.

The linting commands are also now part of our dosh scripts which means
we can run them locally and benefit from the wisdom of the linter when
implementing new features.

Issue #20 Add CI